Clean up index.ts
