+++ parser/src/index.ts
@@ -1,12 +1,11 @@
 import "dotenv/config"
-import { createSpinner } from "nanospinner"
 import { resolve } from "path"
 import { performance } from "perf_hooks"
 import yargsParser from "yargs-parser"
 import { hydrateTreeWithAuthorship } from "./hydrate.js"
-import { getLogLevel, log, LOG_LEVEL } from "./log.js"
+import { log } from "./log.js"
 import { findBranchHead, parseCommit } from "./parse.js"
-import { formatMs, getRepoName, writeRepoToFile } from "./util.js"
+import { describeAsyncJob, formatMs, getRepoName, writeRepoToFile } from "./util.js"
 
 const rawRargs = process.argv.slice(2)
 
@@ -24,66 +23,6 @@ Options:
   --help, -h: Show this help message`)
   process.exit(0)
 }
-const spinner =
-  getLogLevel() <= LOG_LEVEL.INFO
-    ? createSpinner("", {
-        interval: 1000/20,
-        frames: [
-          "                   🚛",
-          "                  🚛 ",
-          "                 🚛  ",
-          "                🚛   ",
-          "               🚛    ",
-          "              🚛     ",
-          "             🚛      ",
-          "            🚛       ",
-          "           🚛        ",
-          "          🚛         ",
-          "         🚛          ",
-          "        🚛           ",
-          "       🚛            ",
-          "      🚛             ",
-          "     🚛              ",
-          "    🚛               ",
-          "   🚛                ",
-          "  🚛                 ",
-          " 🚛                  ",
-          "🚛                   ",
-        ],
-      })
-    : null
-
-async function describeAsyncJob<T>(
-  job: () => Promise<T>,
-  beforeMsg: string,
-  afterMsg: string,
-  errorMsg: string
-) {
-
-    let success = (text: string, final = false) => {
-      if (getLogLevel() === LOG_LEVEL.SILENT) return
-      if (spinner === null) return log.info(text)
-      spinner.success({ text })
-      if (!final) spinner.start()
-    }
-    let error = (text: string) =>
-      spinner === null ? log.error(text) : spinner.error({ text })
-
-    spinner?.start({ text: beforeMsg })
-    try {
-      const startTime = performance.now()
-      const result = await job()
-      const stopTime = performance.now()
-      const suffix = `[${formatMs(stopTime - startTime)}]`
-      success(`${afterMsg} ${suffix}`, true)
-      return result
-    } catch(e) {
-        error(errorMsg)
-        log.error(e)
-        process.exit(1)
-      }
-}
-
 
 const repoDir = args.path ?? "."
 let branch = args.branch ?? null
